Raise error if model max length exceeded
Raise an error if the model max length is exceeded to avoid
uninformative torch `IndexError: index out of range in self` error.

Suggest that the user preprocess the texts to break up long tokens
and/or adjust the config for new models.

Long URLs and texts from other languages without spaces that aren't
supported by the current spacy tokenizer are still going to be a problem
in everyday use.