Use CPU arrays for tokenization output
Tokenizers were called with the option to return PyTorch Tensors.
Since Thinc sets the default tensor type to torch.cuda.FloatTensor
when GPU support is enabled, the tokenization Tensors were allocated
on the GPU. This resulted in many host <-> device transfers during
tokenization and post-processing.

To reduce the number of transfers, this change allocates the
tokenization output as numpy arrays instead. The arrays are only
copied to the GPU when they are used as the input to a
transformer.

Performance of finetuning roberta-base for tagging (RTX 2060 Super):

- Before: ~5800 words per second
- After: ~7350 words per second