Fix compilation with recent Cython 3.x versions
Recent versions of Cython 3.x refuse to compile the
`_get_span2wp_alignment` function because it always requires acquiring
the GIL to check if any exceptions are raised. Change the function
signature to return an `int`, so that `-1` can be used to signal that
an exception occured.

It's not safe to qualify the function `noexcept`, since some calls can
throw an exception.